+++ packages/core/src/js/index.ts
@@ -101,15 +101,14 @@ export type { TimeToDisplayProps, ExpoRouter, ExpoImage, ExpoAsset } from './tra
 
 export { Mask, Unmask } from './replay/CustomMask';
 
+/** @deprecated The `FeedbackButton` component will be removed in a future major version. */
 export { FeedbackButton } from './feedback/FeedbackButton';
 export { FeedbackForm } from './feedback/FeedbackForm';
-export {
-  showFeedbackForm,
-  showFeedbackButton,
-  hideFeedbackButton,
-  enableFeedbackOnShake,
-  disableFeedbackOnShake,
-} from './feedback/FeedbackFormManager';
+export { showFeedbackForm, enableFeedbackOnShake, disableFeedbackOnShake } from './feedback/FeedbackFormManager';
+/** @deprecated `showFeedbackButton` will be removed in a future major version. */
+export { showFeedbackButton } from './feedback/FeedbackFormManager';
+/** @deprecated `hideFeedbackButton` will be removed in a future major version. */
+export { hideFeedbackButton } from './feedback/FeedbackFormManager';
 
 /** @deprecated Use `FeedbackForm` instead. */
 export { FeedbackForm as FeedbackWidget } from './feedback/FeedbackForm';